gcp/setup-rag/requirements.txt
google-cloud-aiplatform>=1.51.0
python-dotenv
